+++ sdk/src/main/scala/com/stratio/sparkta/sdk/TypeOp.scala
@@ -21,6 +21,8 @@ package com.stratio.sparkta.sdk
  */
 
 object TypeOp extends Enumeration {
+
   type TypeOp = Value
-  val Long, Int, String, Double, Boolean, Binary, Date, DateTime, Timestamp, ArrayDouble, ArrayString = Value
+  val BigDecimal, Long, Int, String, Double, Boolean, Binary, Date, DateTime, Timestamp, ArrayDouble,
+  ArrayString = Value
 }